added tamplate for mail sending service
