+++ backend/package.json
@@ -30,12 +30,14 @@
     "@sentry/node": "^7.9.0",
     "@sentry/tracing": "^7.9.0",
     "@types/nodemailer": "^6.4.5",
+    "@types/nodemailer-express-handlebars": "^4.0.2",
     "bcrypt": "^5.0.1",
     "cors": "^2.8.5",
     "express": "^4.18.1",
     "http-status-codes": "^2.2.0",
     "jsonwebtoken": "^8.5.1",
     "nodemailer": "^6.7.7",
+    "nodemailer-express-handlebars": "^5.0.0",
     "passport": "^0.6.0",
     "passport-local": "^1.0.0",
     "yup": "^0.32.11"
