+++ backend/src/services/mail_verification/send_activation_link/constants.ts
