+++ package.json
@@ -25,7 +25,8 @@
     "build": "yarn build:shared && yarn build:frontend && yarn build:backend"
   },
   "dependencies": {
-    "@sentry/node": "^7.9.0"
+    "@sentry/node": "^7.9.0",
+    "@types/nodemailer-express-handlebars": "^4.0.2"
   },
   "devDependencies": {
     "@ls-lint/ls-lint": "1.11.2",
